feat(client): useQuery 적용 부분 컴포넌트화
