+++ apps/client/src/app/page.tsx
@@ -6,16 +6,13 @@ import { useRouter } from 'next/navigation';
 import { Suspense } from 'react';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
-import JobPostingCard from '@/components/JobPostingCard';
+import JobPostingList from '@/components/JobPostingList';
 import LoginBox from '@/components/LoginBox';
 import MouCompanyCard from '@/components/MouCompanyCard';
 import { POSITION_LIST } from '@/constants/common';
-import { useGetJobPostingList } from '@/hooks/api/job-posting/useGetJobPostingList';
 
 const MainPage = () => {
   const router = useRouter();
-
-  const { jobPostingListData } = useGetJobPostingList();
 
   const handleGoFullViewJobPosting = () => {
     window.open(
@@ -118,18 +115,9 @@ const MainPage = () => {
               </Text>
               <Link onClick={handleGoFullViewJobPosting}>전체 보기</Link>
             </Flex>
-            <Grid templateColumns="repeat(3, 1fr)" gap="32px">
-              <Suspense fallback={<Spinner color="priamry" />}>
-                {jobPostingListData.map((jobPosting) => (
-                  <JobPostingCard
-                    title={jobPosting.title}
-                    imageUrl={jobPosting.imageUrl}
-                    companyName={jobPosting.companyName}
-                    detailLink={jobPosting.detailLink}
-                  />
-                ))}
-              </Suspense>
-            </Grid>
+            <Suspense fallback={<Spinner color="priamry" />}>
+              <JobPostingList />
+            </Suspense>
           </Flex>
         </Box>
       </Box>
